stash all changes

+++ contracts/interfaces/IERC721MintableBurnable.sol
@@ -1,12 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-import "@openzeppelin/contracts/token/ERC721/IERC721.sol";
-import "@openzeppelin/contracts/token/ERC721/extensions/IERC721Enumerable.sol";
-import "@openzeppelin/contracts/token/ERC721/extensions/IERC721Metadata.sol";
+import {IERC721} from "@openzeppelin/contracts/token/ERC721/IERC721.sol";
 
 interface IERC721MintableBurnable is IERC721 {
     function safeMint(address account, uint256 tokenId, string memory uri) external;
 
-    function burnFrom(address account, uint tokenId) external;
+    function burnFrom(address account, uint256 tokenId) external;
 }
